+++ packages/jdm-editor/src/components/code-editor/index.ts
@@ -1,1 +1,2 @@
-export * from './ce';
+export { CodeEditor, type CodeEditorRef, type CodeEditorProps } from './ce';
+export { ExpressionBuilder, type ExpressionBuilderProps, type ExpressionBuilderRef } from './business';