+++ packages/jdm-editor/src/components/decision-graph/dg-infer.tsx
@@ -139,9 +139,7 @@ type InferNodeTypes = (
 const inferNodeTypes: InferNodeTypes = ({ decisionGraph, nodeTypes, customNodes }, prevState, graphWalker) => {
   let isModified = false;
   const newNodeTypes = produce(nodeTypes, (draft) => {
-    const graphExecResults = graphWalker.walk(decisionGraph);
-    if (Array.isArray(graphExecResults)) {
-    graphExecResults.forEach(({ node, incomers }) => {
+    for (const { node, incomers } of graphWalker.walk(decisionGraph)) {
       if (node.type === 'inputNode') {
         return;
       }
@@ -196,9 +194,8 @@ const inferNodeTypes: InferNodeTypes = ({ decisionGraph, nodeTypes, customNodes 
         draft[node.id] ??= {};
         draft[node.id][NodeTypeKind.InferredOutput] = inferredOutputType;
       }
-    });
-  }
-});
+    }
+  });
 
   return {
     isModified,