+++ packages/jdm-editor/src/components/decision-graph/nodes/graph-node.tsx
@@ -76,7 +76,7 @@ export const GraphNode = React.forwardRef<HTMLDivElement, GraphNodeProps>(
       // !displayError && {
       //   key: 'copy-clipboard',
       //   label: <SpacedText left='复制' right={platform.shortcut('Ctrl + C')} />,
-      //   onClick: () => {},
+      //   onClick: () => graphActions.copyNodes([id]),
       // },
       !displayError && {
         key: 'duplicate',