+++ packages/jdm-editor/src/components/decision-graph/simulator/simulation.types.ts
@@ -36,7 +36,7 @@ export type SimulationTrace<Trace = TraceDataVariants> = {
   name: string;
   id: string;
   performance: string | null;
-  traceData: Trace;
+  traceData: Trace | null;
   order?: number;
 };
 
@@ -46,7 +46,7 @@ export type SimulationTraceDataFunction = {
 
 type Arrayable<T> = T | T[];
 
-export type SimulationTraceDataExpression = Arrayable<Record<string, { result: string }>>;
+export type SimulationTraceDataExpression = Arrayable<Record<string, { result: unknown }>>;
 
 type SimulationTraceDataTableSingle = {
   index: number;